+++ src/controllers/SourcesController.php
@@ -583,7 +583,7 @@ class SourcesController extends Controller
             $identifierOptions = [];
 
             foreach ($craftFields as $field) {
-                if (in_array($field['type'], ['field']) && $field['supported'] && !in_array($field['handle'], ['variantTitle', 'sku', 'price'])) {
+                if (in_array($field['type'], ['field']) && $field['supported'] && !in_array($field['handle'], ['variantTitle', 'sku', 'price', 'author', 'postDate', 'expiryDate', 'enabled'])) {
                     $identifierOptions[] = [
                         'label' => $field['name'],
                         'value' => $field['handle'],